feat: restructure test files
