+++ src/ratelimit/mod.rs
@@ -34,81 +34,3 @@ impl RatelimitConfiguration {
 		Self { redis, cache, prefix }
 	}
 }
-
-#[cfg(test)]
-mod tests {
-
-	use self::single::{FixedWindow, SlidingWindow, TokenBucket};
-
-	use super::*;
-	use dotenv::dotenv;
-
-	#[tokio::test]
-	async fn test_fixed_window() {
-		dotenv().ok();
-
-		let connection_str = std::env::var("UPSTASH_REDIS_URL").unwrap_or_else(|_| panic!("Expecting UPSTASH_REDIS_URL to be set"));
-		let Ok(redis) = redis::Client::open(connection_str) else {
-			panic!("Failed to connect")
-		};
-		let client = RatelimitConfiguration::new(redis, true, None);
-
-		let identifier = "anonymous32";
-
-		let ratelimit = FixedWindow::new(client, 10, "60s");
-
-		for _ in 1..11 {
-			let res = ratelimit.limit(identifier, None).await;
-
-			assert!(res.success);
-		}
-
-		let res = ratelimit.limit(identifier, None).await;
-		assert!(!res.success);
-	}
-	#[tokio::test]
-	async fn test_sliding_window() {
-		dotenv().ok();
-
-		let connection_str = std::env::var("UPSTASH_REDIS_URL").unwrap_or_else(|_| panic!("Expecting UPSTASH_REDIS_URL to be set"));
-		let Ok(redis) = redis::Client::open(connection_str) else {
-			panic!("Failed to connect")
-		};
-		let client = RatelimitConfiguration::new(redis, true, None);
-
-		let ratelimit = SlidingWindow::new(client, 10, "60s");
-
-		let identifier = "anonymous12";
-
-		for _ in 1..11 {
-			let res = ratelimit.limit(identifier, None).await;
-			assert!(res.success);
-		}
-
-		let res = ratelimit.limit(identifier, None).await;
-		assert!(!res.success);
-	}
-
-	#[tokio::test]
-	async fn test_token_window() {
-		dotenv().ok();
-
-		let connection_str = std::env::var("UPSTASH_REDIS_URL").unwrap_or_else(|_| panic!("Expecting UPSTASH_REDIS_URL to be set"));
-		let Ok(redis) = redis::Client::open(connection_str) else {
-			panic!("Failed to connect")
-		};
-		let client = RatelimitConfiguration::new(redis, true, None);
-
-		let ratelimit = TokenBucket::new(client, 10, "10s", 5);
-
-		let identifier = "anonymous52";
-
-		for _ in 1..11 {
-			let res = ratelimit.limit(identifier, None).await;
-			assert!(res.success);
-		}
-
-		let res = ratelimit.limit(identifier, None).await;
-		assert!(!res.success);
-	}
-}
